add more staff

+++ Doctor.cpp
@@ -2,9 +2,9 @@
 Doctor::Doctor() {
 	Department = 'NULL';
 	Specialization = 'NULL';
-	Rank = 0.0;
+	Rank = 0;
 }
-Doctor::Doctor( string Id, string fName, string lName, string Gender, int Age, string mobNum, string add, string department, string specialization,double rank) :person(Id, fName, lName, Gender, Age, mobNum, add) {
+Doctor::Doctor( string Id, string fName, string lName, string Gender, int Age, string mobNum, string add, string department, string specialization,int rank) :person(Id, fName, lName, Gender, Age, mobNum, add) {
 	Department = department;
 	Specialization = specialization;
 	Rank = rank;
